refactor: Update treatment schedule dates and time in allure_dev_data.sql
Update the dates and time for a treatment schedule in the allure_dev_data.sql file. The previous dates and time were changed to '2024-11-01 10:00:00' and '2024-11-01 11:00:00' respectively.

refactor: Update NotificationBar in CustomersView.vue

Update the NotificationBar component in CustomersView.vue to conditionally render it based on the number of upcoming birthdays. If there are upcoming birthdays, the NotificationBar will be displayed with the 'info' color and the mdiCalendarAccount icon. Otherwise, it will not be rendered.